+++ addons/binding/org.openhab.binding.plclogo/META-INF/MANIFEST.MF
@@ -9,8 +9,6 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ClassPath: .,
   lib/Moka7-1.0.2_io_patch.jar
 Bundle-ActivationPolicy: lazy
-Export-Package: org.openhab.binding.plclogo,
- org.openhab.binding.plclogo.handler
 Import-Package: org.apache.commons.net.util,
  org.eclipse.jdt.annotation;resolution:=optional,
  org.eclipse.smarthome.config.core,
@@ -24,7 +22,5 @@ Import-Package: org.apache.commons.net.util,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
  org.eclipse.smarthome.model.script.actions,
- org.openhab.binding.plclogo,
- org.openhab.binding.plclogo.handler,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
