+++ addons/binding/org.openhab.binding.plclogo/src/main/java/org/openhab/binding/plclogo/internal/config/PLCDigitalConfiguration.java
@@ -12,7 +12,7 @@
  */
 package org.openhab.binding.plclogo.internal.config;
 
-import static org.openhab.binding.plclogo.PLCLogoBindingConstants.*;
+import static org.openhab.binding.plclogo.internal.PLCLogoBindingConstants.*;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
